Add hard mining strategies
