+++ bin/train_mixed.sh
@@ -7,6 +7,7 @@ do
                 --model_name_or_path "roberta-base" \
                 --tokenizer "roberta-base" \
                 --loss_fct "online-constrastive" \
+                --sample_selection "both" \
                 --distance "cosine" \
                 --directory "kpm_6_folds/fold_$fold_id" \
                 --logging_dir "runs/mixed/fold_$fold_id" \
