+++ src/losses/__init__.py
@@ -1,0 +1,8 @@
+from .contrastive_loss import ContrastiveLoss
+from .cosine_similarity_loss import CosineSimilarityLoss
+from .online_constrastive_loss import OnlineContrastiveLoss
+from .online_triplet_loss import (
+    HardNegativeTripletLoss,
+    HardPositiveTripletLoss,
+)
+from .triplet_loss import TripletLoss
