+++ src/mixed/model_argument.py
@@ -34,9 +34,10 @@ class ModelArguments:
     stance_dim: int = field(default=32, metadata={"help": "Hidden representation dimension used for encoding stance."})
     text_dim: int = field(default=32, metadata={"help": "Hidden representation dimension used for encoding text."})
     drop_rate: float = field(default=0.1, metadata={"help": "Model dropout rate."})
-    loss_fct: str = field(default="constrastive", metadata={"help": "Loss function used for training siamese network."})
+    loss_fct: str = field(default="constrastive", metadata={"help": "Loss function used for constrastive network."})
+    sample_selection: str = field(default="all", metadata={"help": "Online triplet selection strategy."})
     distance: str = field(
         default="euclidean", metadata={"help": "Function that returns a distance between two emeddings."}
     )
-    pair_margin: float = field(default=1.0, metadata={"help": "Margin distance value for pair loss."})
-    triplet_margin: float = field(default=1.0, metadata={"help": "Margin distance value for triplet loss."})
+    pair_margin: float = field(default=0.5, metadata={"help": "Margin distance value for pair loss."})
+    triplet_margin: float = field(default=0.5, metadata={"help": "Margin distance value for triplet loss."})
