+++ src/question_answering/models.py
@@ -3,13 +3,13 @@ import torch.nn as nn
 import torch.nn.functional as F
 from transformers import AutoConfig, AutoModel
 
-from src.question_answering.model_argument import ModelArguments
-from src.train_utils.distance import SiameseDistanceMetric
-from src.train_utils.losses import (
+from src.losses import (
     ContrastiveLoss,
     CosineSimilarityLoss,
     OnlineContrastiveLoss,
 )
+from src.question_answering.model_argument import ModelArguments
+from src.train_utils.distance import SiameseDistanceMetric
 from src.utils.logging import custom_logger
 
 logger = custom_logger(__name__)
